move build tests out of platform tests
