+++ packages/adapter-netlify/package.json
@@ -38,7 +38,7 @@
 		"lint": "oxfmt --check .",
 		"format": "oxfmt --write .",
 		"test:unit": "vitest run",
-		"test:build": "pnpm -r --workspace-concurrency 1 --filter=\"./test/**\" build",
+		"test:build": "pnpm -r --workspace-concurrency 1 --filter=\"./test/**\" test:build",
 		"test": "pnpm test:unit && pnpm test:build"
 	},
 	"dependencies": {
