+++ packages/adapter-netlify/test/apps/basic/test/test.js
@@ -22,16 +22,3 @@ test('read from $app/server works', async ({ request }) => {
 	const response = await request.get('/read');
 	expect(await response.text()).toBe(content);
 });
-
-test('_redirects are copied to publish directory', () => {
-	const redirects = fs.readFileSync(
-		path.resolve(import.meta.dirname, '../build/_redirects'),
-		'utf-8'
-	);
-	expect(redirects).toContain('/redirect-me /greeting/redirected 301');
-});
-
-test('_headers are copied to publish directory', () => {
-	const headers = fs.readFileSync(path.resolve(import.meta.dirname, '../build/_headers'), 'utf-8');
-	expect(headers).toContain('X-Custom-Header: test-value');
-});
