+++ packages/adapter-netlify/test/apps/edge/build.test.js
@@ -1,0 +1,8 @@
+import fs from 'node:fs';
+import path from 'node:path';
+import { test, expect } from 'vitest';
+
+test('_headers are copied to publish directory', () => {
+	const headers = fs.readFileSync(path.resolve(import.meta.dirname, './build/_headers'), 'utf-8');
+	expect(headers).toContain('X-Custom-Header: test-value');
+});
