+++ packages/adapter-netlify/test/apps/split/build.test.js
@@ -1,0 +1,11 @@
+import fs from 'node:fs';
+import path from 'node:path';
+import { test, expect } from 'vitest';
+
+test('_redirects are copied to publish directory', () => {
+	const redirects = fs.readFileSync(
+		path.resolve(import.meta.dirname, './build/_redirects'),
+		'utf-8'
+	);
+	expect(redirects).toContain('/redirect-me /greeting/redirected 301');
+});
