+++ packages/adapter-netlify/vitest.config.js
@@ -2,4 +2,8 @@
 
 import { defineConfig } from 'vitest/config';
 
-export default defineConfig({});
+export default defineConfig({
+	test: {
+		include: ['*.spec.{js,ts}'],
+	}
+});
